[RN] Implement Labels on mobile
